move category list to separated file
