+++ lib/features/home/presentation/widgets/list_of_category.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+import 'custom_category_circle.dart';
+
+class ListOfCategory extends StatelessWidget {
+  const ListOfCategory({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 60,
+      child: ListView.builder(
+        scrollDirection: Axis.horizontal,
+        itemBuilder: (context, index) {
+          return const CustomCategoryCircle();
+        },
+      ),
+    );
+  }
+}
