+++ pubspec.yaml
@@ -52,6 +52,8 @@ dependencies:
   get_it: ^7.7.0
   # Flutter plugin for selecting images from the Android and iOS image library
   image_picker: ^1.1.2
+  # Provides a collection of Flutter grids layouts
+  flutter_staggered_grid_view: ^0.7.0
   # Networking
   dio: ^5.4.3+1
   pretty_dio_logger: ^1.3.1
